+++ publish/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
+## [5.0.30] - 2025-01-29
+
+### Fixed
+- 🔧 Added global exception handlers to updater - will now ALWAYS create log files even if crashes
+- 🔧 Moved logging to App.xaml.cs to catch constructor and InitializeComponent errors
+
 ## [5.0.29] - 2025-01-29
 
 ### Fixed